chore(widget-container): rename
