+++ src/services/dashboards/dashboard-detail/modules/DashboardWidgetContainer.vue
